requirements.txt
discord.py
pymongo
openai
tiktoken
python-dotenv
lmstudio
pydantic